python-services/requirements-test.txt
# Testing dependencies for K-ERP Python services

# Test framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.5.0

# Mock and fixtures
faker>=22.0.0
factory-boy>=3.3.0

# gRPC testing
grpcio-testing>=1.60.0

# HTTP mocking
responses>=0.24.0
httpx>=0.26.0

# Type checking
mypy>=1.8.0
types-protobuf>=4.24.0

# Code quality
ruff>=0.1.0
black>=24.0.0
isort>=5.13.0
